+++ demo/src/main/java/com/example/demo/service/ClienteService.java
@@ -7,10 +7,10 @@ import com.example.demo.entities.Cliente;
 public interface ClienteService { 
     
     Collection<Cliente> obtenerTodos(); //Obtenemos todos los clientes que existan
-    Cliente obtenerPorId(Integer id); //Obtenemos a un solo cliente por su id 
+    Cliente obtenerPorId(Long id); //Obtenemos a un solo cliente por su id 
 
     
     Cliente guardar(Cliente cliente); //Guardamos un objeto cliente o tambien se puede actualizar
-    void eliminar(Integer id); //Eliminar un cliente por su id
+    void eliminar(Long id); //Eliminar un cliente por su id
     boolean existe(Integer id); //Verificar si existe algun cliente con ese ID
 }
