+++ demo/src/main/java/com/example/demo/service/ProductoService.java
@@ -7,10 +7,10 @@ import com.example.demo.entities.Producto;
 public interface ProductoService { // Interfaz, los metodos que debe tener el servicio de producto
     
     Collection<Producto> obtenerTodos(); //Obtener todos los productos
-    Producto obtenerPorId(Integer id); //Obtener solo uno en especifico mediante su id
+    Producto obtenerPorId(Long id); //Obtener solo uno en especifico mediante su id
 
     
     Producto guardar(Producto producto); //Guardar o actualizar un producto
-    void eliminar(Integer id); //Eliminar un producto
-    boolean existe(Integer id); //Verificar si el producto existe
+    void eliminar(Long id); //Eliminar un producto
+    boolean existe(Long id); //Verificar si el producto existe
 }
